pyproject.toml
[tool.poetry]
name = "klash"
version = "2.0"
description = "SMM usecase to social networks"
authors = ["burchalovskiy@gmail.com"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = "*"
loguru = "*"
python-dotenv = "*"
rabbitmq = "*"
environs = "*"
uvloop = "*"
tenacity = "*"
APScheduler = "*"
rsa = "*"
aiohttp = "*"
asyncpg = "*"
aerich = "*"
fastapi-admin = { git = 'https://github.com/fastapi-admin/fastapi-admin.git' }
instagrapi = "^1.16.30"
telebot = "^0.0.4"
vk-api = "^11.9.9"
redis = "^4.4.2"
asyncio = "^3.4.3"
tortoise-orm = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-httpx = "*"
pytest-mock = "*"
coverage = "*"
fakeredis = "*"
pytest-cov = "*"
pytest-socket = "*"
pytest-asyncio = "*"
mypy = "*"
black = "*"
safety = "*"
taskipy = "*"
isort = "*"

[tool.taskipy.tasks]
isort = "isort app"
lint = "task isort-lint"
isort-lint = "task isort -c"

tests = "pytest tests"

black = "black app/"
black-lint = "black app/ --check"

mypy-lint = "mypy --install-types --non-interactive app/"

[tool.isort]
# isort configuration:
# https://github.com/timothycrosley/isort/wiki/isort-Settings
include_trailing_comma = true
use_parentheses = true
# See https://github.com/timothycrosley/isort#multi-line-output-modes
multi_line_output = 3
line_length = 120
known_first_party = "app"
default_section = "THIRDPARTY"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"



[tool.black]
line-length = 120
target-version = ['py310']
extend-exclude = '''
(
  /(
    # Exclude folders
    | app/migrations
  )/
  # Eclude files
)
'''
skip-string-normalization = true


[tool.mypy]
# Mypy configuration:
# https://mypy.readthedocs.io/en/latest/config_file.html
allow_redefinition = false
check_untyped_defs = true
disallow_untyped_decorators = false
disallow_any_generics = false
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = true
local_partial_types = true
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = false
pretty = true
follow_imports = "silent"
follow_imports_for_stubs = true
show_error_codes = true
plugins = "pydantic.mypy"

exclude = [
    'test_',
    'models\.py'
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "html.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "docker.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.migrations.*"
ignore_errors = true



[tool.aerich]
tortoise_orm = "app.database.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
